+++ lib/src/rankings/bloc/event.dart
@@ -1,0 +1,22 @@
+part of 'bloc.dart';
+
+sealed class RankingsEvent {
+  const RankingsEvent();
+}
+
+final class RankingsStarted extends RankingsEvent {
+  const RankingsStarted();
+}
+
+final class RankingsTypeSelected extends RankingsEvent {
+  const RankingsTypeSelected(this.value);
+  final RankingsType value;
+}
+
+final class RankingsRefreshRequested extends RankingsEvent {
+  const RankingsRefreshRequested();
+}
+
+final class RankingsMoreRequested extends RankingsEvent {
+  const RankingsMoreRequested();
+}